Move buildifier_utils to buildifier/utils
